Move core marin integration to core
