+++ structuralcodes/sections/section_integrators/__init__.py
@@ -1,8 +1,10 @@
 """Classes for integrating the response of sections."""
 
+from structuralcodes.core._marin_integration import marin_integration
+
 from ._factory import integrator_factory
 from ._fiber_integrator import FiberIntegrator
-from ._marin_integrator import MarinIntegrator, marin_integration
+from ._marin_integrator import MarinIntegrator
 from ._section_integrator import SectionIntegrator
 
 __all__ = [
